finish database and session template

+++ Spring/Lec 5 Spring boot/Spring/src/main/java/com/spring/boot/service/impl/TeacherServiceImpl.java
@@ -1,5 +1,6 @@
 package com.spring.boot.service.impl;
 
+import com.spring.boot.dto.TeacherDto;
 import com.spring.boot.model.Teacher;
 import com.spring.boot.repo.TeacherRepo;
 import com.spring.boot.service.TeacherService;
@@ -7,6 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class TeacherServiceImpl implements TeacherService {
@@ -19,27 +23,70 @@ public class TeacherServiceImpl implements TeacherService {
     }
 
     @Override
-    public List<Teacher> getAllTeachers() {
-        return teacherRepo.findAll();
+    public List<TeacherDto> getAllTeachers() {
+        List<Teacher> teachers = teacherRepo.findAll();
+
+        return teachers.stream().map(teacher ->
+                new TeacherDto(teacher.getId(), teacher.getUsername(), teacher.getPassword()))
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Teacher addTeacher(Teacher teacher) {
-        return teacherRepo.save(teacher);
+    public TeacherDto addTeacher(TeacherDto teacherDto) {
+        if(Objects.nonNull(teacherDto.getId())){
+            throw new RuntimeException("id must be null");
+        }
+        if(Objects.isNull(teacherDto.getUsername())){
+            throw new RuntimeException("user name must be not null");
+        }
+        if(Objects.isNull(teacherDto.getPassword())){
+            throw new RuntimeException("password must be not null");
+        }
+        Optional<Teacher> teacherOptional = teacherRepo.findByusername(teacherDto.getUsername());
+        if(teacherOptional.isPresent()){
+            throw new RuntimeException("teacher is already exist with same username");
+        }
+        Teacher teacher = teacherRepo.save(new Teacher(teacherDto.getUsername(),teacherDto.getPassword()));
+        teacherDto.setId(teacher.getId());
+        return teacherDto;
     }
 
     @Override
-    public Teacher updateTeacher(Teacher teacher) {
-        return teacherRepo.save(teacher);
+    public TeacherDto updateTeacher(TeacherDto teacherDto) {
+        if(Objects.isNull(teacherDto.getId())){
+            throw new RuntimeException("id must be not null");
+        }
+        Teacher teacher = teacherRepo.save(new Teacher(teacherDto.getId(),teacherDto.getUsername(),teacherDto.getPassword()));
+        return teacherDto;
+
     }
 
     @Override
     public void deleteTeacher(Long id) {
+        Optional<Teacher> teacherOptional = teacherRepo.findById(id);
+        if(teacherOptional.isEmpty()){
+            throw new RuntimeException("teacher not found");
+        }
         teacherRepo.deleteById(id);
     }
 
     @Override
-    public Teacher getTeacherById(Long id) {
-        return teacherRepo.findById(id).get();
+    public TeacherDto getTeacherById(Long id) {
+        Optional<Teacher> teacherOptional = teacherRepo.findById(id);
+        if(teacherOptional.isEmpty()){
+            throw new RuntimeException("teacher not found");
+        }
+        Teacher teacher =teacherOptional.get();
+        return new TeacherDto(teacher.getId(), teacher.getUsername(), teacher.getPassword());
+    }
+
+    @Override
+    public TeacherDto getTeacherByUserName(String userName) {
+        Optional<Teacher> teacherOptional = teacherRepo.findByusername(userName);
+        if(teacherOptional.isEmpty()){
+            throw new RuntimeException("teacher not found");
+        }
+        Teacher teacher =teacherOptional.get();
+        return new TeacherDto(teacher.getId(), teacher.getUsername(), teacher.getPassword());
     }
 }
